interview_ai/requirements.txt
fastapi
uvicorn
deepface
numpy
pillow
tensorflow==2.15.0
mtcnn
retina-face
gunicorn
python-dotenv